dom/os/test/test_writefile.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>os write test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141021">os tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  var jsFuns = SpecialPowers.Cu.getJSTestingFunctions();

  function runTest(expected) {
    return new Promise((res, rej) => {
      var workerCode = `
        try {
          var path = navigator.os.TEMP_DIR + '/test_writefile';
          var file = navigator.os.open(path,
            navigator.os.RDWR | navigator.os.CREAT, 0666);
          var arr = new Uint8Array("` + expected + `".split('').map(c => c.charCodeAt(0)));
          var written = navigator.os.write(file, arr);
          var close1 = navigator.os.close(file);

          file = navigator.os.open(path, navigator.os.RDONLY, navigator.os.IREAD);
          var data = navigator.os.read(file, 200);
          var close2 = navigator.os.close(file);

          postMessage(JSON.stringify({
            data: data.join(','),
            written: written,
            close1: close1,
            close2: close2
          }));
        }
        catch (ex) {
          postMessage("write err: " + ex);
        }
        `;

      var workerBlob = new Blob([workerCode], { type: 'application/javascript' });
      var w = new Worker(URL.createObjectURL(workerBlob));

      w.onmessage = e => {
        ok(e.data.indexOf('write err') === -1, 'ran successfully');
        if (e.data.indexOf('write err') !== -1) {
          dump(e.data + '\n');
          return res();
        }

        try {
          var data = JSON.parse(e.data);
        }
        catch(ex) {
          ok(null, ex, fn + ' is valid JSON');
        }

        ok(data.written === expected.length, 'written. exp: ' + expected.length + ', was: ' + data.written);
        var expectedData = expected.split('').map(c=>c.charCodeAt(0)).join(',');
        if (data.data !== expectedData) {
          dump('expected "' + expectedData + '", but was "' + data.data + '"\n');
        }
        ok(data.data === expectedData, 'data');
        ok(data.close1 === 0, 'close1');
        ok(data.close2 === 0, 'close2');

        res();
      };
    });
  }

  SimpleTest.waitForExplicitFinish();

  Promise.all([
    runTest('Hello world')
  ]).then(() => {
    SimpleTest.finish();
  });
  </script>
</body>
</html>
